Email includes link to dataset and dataset name
