+++ utils.py
@@ -4,6 +4,24 @@ from datetime import datetime
 import re
 from elasticsearch_dsl import Search, Q
 from models import get_tweets_index_name
+from celery import Celery
+
+def make_celery(app):
+    '''As described in Flask documentation. Adds a Flask application context to the Celery worker thread.'''
+    celery = Celery(
+        app.name,
+        backend=app.config['CELERY_RESULT_BACKEND'],
+        broker=app.config['CELERY_BROKER_URL']
+    )
+    celery.conf.update(app.config)
+
+    class ContextTask(celery.Task):
+        def __call__(self, *args, **kwargs):
+            with app.app_context():
+                return self.run(*args, **kwargs)
+
+    celery.Task = ContextTask
+    return celery
 
 
 def write_json(filepath, obj):
